requirements.txt
pypdf2
pillow
python-dotenv